+++ src/main/java/tracer/motionProfiles/ConstantVelocityProfile.java
@@ -1,6 +1,7 @@
-package tracer;
+package tracer.motionProfiles;
 
 import com.flash3388.flashlib.time.Time;
+import tracer.MotionParameters;
 
 public class ConstantVelocityProfile extends Profile {
     private final double maxVelocity;
