+++ src/main/java/tracer/motionProfiles/ConvexProfile.java
@@ -1,6 +1,7 @@
-package tracer;
+package tracer.motionProfiles;
 
 import com.flash3388.flashlib.time.Time;
+import tracer.MotionParameters;
 
 public class ConvexProfile extends Profile {
     private final double maxAcceleration;
